% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_BinCont_copula.R
\name{binary_continuous_loglik}
\alias{binary_continuous_loglik}
\title{Title}
\usage{
binary_continuous_loglik(
  para,
  X,
  Y,
  copula_family,
  marginal_surrogate,
  marginal_true
)
}
\arguments{
\item{marginal_true}{}
}
\description{
Title
}
