% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_BinCont_copula.R
\name{cdf_fun}
\alias{cdf_fun}
\title{Function factory for distribution functions}
\usage{
cdf_fun(x, mean, extra_par, family)
}
\arguments{
\item{mean}{Mean parameter.}

\item{extra_par}{Additional parameters. The specific meaning of these extra
parameters depends on \code{family}.}

\item{family}{Distributional family, one of the following:
\itemize{
\item \code{"normal"}: normal distribution where \code{mean} is the mean and \code{extra_par} is
the standard deviation.
\item \code{"logistic"}: logistic distribution as parameterized in \code{stats::plogis()} where
\code{mean} and \code{extra_par} correspond to \code{location} and \code{scale}, respectively.
\item \code{"t"}: t distribution as parameterized in \code{stats::pt()} where \code{mean} and
\code{extra_par} correspond to \code{ncp} and \code{df}, respectively.
}}
}
\value{
A distribution function that has a single argument. This is the
vector of values in which the distribution function is evaluated.
}
\description{
Function factory for distribution functions
}
