% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{clayton_loglik_copula_scale}
\alias{clayton_loglik_copula_scale}
\title{Loglikelihood on the Copula Scale for the Clayton Copula}
\usage{
clayton_loglik_copula_scale(theta, u, v, d1, d2)
}
\arguments{
\item{theta}{Copula parameter}

\item{u}{A numeric vector. Corresponds to first variable on the copula scale.}

\item{v}{A numeric vector. Corresponds to second variable on the copula
scale.}

\item{d1}{An integer vector. Indicates whether first variable is observed or
 right-censored,
* `d1[i] = 1` if `u[i]` corresponds to non-censored value
* `d1[i] = 0` if `u[i]` corresponds to right-censored value}

\item{d2}{An integer vector. Indicates whether first variable is observed or
 right-censored,
* `d1[i] = 1` if `u[i]` corresponds to non-censored value
* `d1[i] = 0` if `u[i]` corresponds to right-censored value}
}
\value{
Value of the copula loglikelihood evaluated in `theta`.
}
\description{
Loglikelihood on the Copula Scale for the Clayton Copula
}
