% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA.BinCont.R
\name{compute_mutinf_BinCont_mixture}
\alias{compute_mutinf_BinCont_mixture}
\title{Computes the mutual information in the binary-continous setting for the
normal-mixture model.}
\usage{
compute_mutinf_BinCont_mixture(mu_s, sigma_s, pi_hier, method = "Monte Carlo")
}
\arguments{
\item{mu_s}{Vector of means of \eqn{\Delta S} in the four components:
\eqn{(\mu_{ \ast 00}, \mu_{ \ast 01}, \mu_{ \ast 10}, \mu_{ \ast 11})}}

\item{sigma_s}{Vector of variances of \eqn{\Delta S} in the four components:
\eqn{(\sigma_{ \ast 00}, \sigma_{ \ast 01}, \sigma_{ \ast 10}, \sigma_{ \ast 11})}}

\item{pi_hier}{Vector of probabilities for the four components:
\eqn{(\pi_{00}, \pi_{01}, \pi_{10}, \pi_{11})}}

\item{method}{Method to compute the mutual information, can be:
\itemize{
\item \code{"Monte Carlo"} (default): The integrals are computed through Monte Carlo
integration.
\item \code{"Numerical"}: The integrals are computed through numerical integration.
}}
}
\value{
(numeric) the mutual information
}
\description{
The \code{\link[=compute_mutinf_BinCont_mixture]{compute_mutinf_BinCont_mixture()}} function computes the mutual
information between \eqn{\Delta S} and \eqn{\Delta T} for the normal-mixture
model where \eqn{S} is continuous and \eqn{T} is binary.
}
