% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITCI-Dvine-inference-utils.R
\name{delta_method_log_mutinfo}
\alias{delta_method_log_mutinfo}
\title{Variance of log-mutual information based on the delta method}
\usage{
delta_method_log_mutinfo(
  fitted_model,
  copula_par_unid,
  copula_family2,
  rotation_par_unid,
  n_prec,
  mutinfo_estimator = NULL,
  composite,
  seed,
  eps = 0.001
)
}
\arguments{
\item{eps}{(numeric) Step size for finite difference in numeric
differentiation}
}
\value{
(numeric) Variance for the estimated ICA based on the delta method,
holding the unidentifiable parameters fixed at the user supplied values.
}
\description{
\code{\link[=delta_method_log_mutinfo]{delta_method_log_mutinfo()}} computes the variance of the estimated log
mutual information, given the unidentifiable parameters.
}
