% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_ContCont_copula.R
\name{estimate_ICA_ContCont}
\alias{estimate_ICA_ContCont}
\title{Estimate ICA in Ordinal-Ordinal Setting}
\usage{
estimate_ICA_ContCont(delta_S, delta_T)
}
\arguments{
\item{delta_S}{(numeric) Vector of individual causal treatment effects on the
surrogate.}

\item{delta_T}{(numeric) Vector of individual causal treatment effects on the
true endpoint.}
}
\value{
(numeric) Estimated ICA
}
\description{
\code{estimate_ICA_ContCont()} estimates the individual causal association (ICA) for
a sample of individual causal treatment effects with a continuous surrogate and
true endpoint. The ICA in this setting is defined as the squared informational
coefficient of correlation, which is a transformation of the mutual information.
The mutual information is estimated with \code{fnn::mutinfo()}.
}
