% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{estimate_marginal}
\alias{estimate_marginal}
\title{Estimate marginal distribution using ML}
\usage{
estimate_marginal(Y, marginal_Y, starting_values)
}
\value{
Estimated parameters
}
\description{
Estimate marginal distribution using ML
}
