% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_ContCont_copula.R
\name{fit_copula_ContCont}
\alias{fit_copula_ContCont}
\title{Fit continuous-continuous vine copula model}
\usage{
fit_copula_ContCont(
  data,
  copula_family,
  marginal_S0,
  marginal_S1,
  marginal_T0,
  marginal_T1,
  start_copula,
  method = "BFGS",
  maxit = 500
)
}
\arguments{
\item{data}{A data frame in the correct format (See details).}

\item{copula_family}{One of the following parametric copula families:
\code{"clayton"}, \code{"frank"}, \code{"gaussian"}, or \code{"gumbel"}. The first element in
\code{copula_family} corresponds to the control group, the second to the
experimental group.}

\item{method}{Optimization algorithm for maximizing the objective function.
For all options, see \code{?maxLik::maxLik}. Defaults to \code{"BFGRS"}.}

\item{maxit}{Maximum number of iterations for the numeric optimization,
defaults to 500.}
}
\value{
Returns an S3 object that can be used to perform the sensitivity
analysis with \code{\link[=sensitivity_analysis_copula]{sensitivity_analysis_copula()}}.
}
\description{
Fit continuous-continuous vine copula model
}
\seealso{
\code{\link[=sensitivity_analysis_copula]{sensitivity_analysis_copula()}}
}
\author{
Florian Stijven
}
