% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA.BinCont.R
\name{fit_mixture_BinCont}
\alias{fit_mixture_BinCont}
\title{Fit normal mixture distribution given the component probabilities.}
\usage{
fit_mixture_BinCont(data_S, theta_start_S, pi_hier, Diff.Sigma)
}
\arguments{
\item{data_S}{Vector of potential outcomes, either \eqn{S_0} or \eqn{S_1}.}

\item{theta_start_S}{Starting values.}

\item{pi_hier}{Vector of probabilities for the four components:
\eqn{(\pi_{00}, \pi_{01}, \pi_{10}, \pi_{11})}. These are fixed during
estimation.}

\item{Diff.Sigma}{Are all component variances assumed equal?
\itemize{
\item \code{TRUE}: all component variances within treatment group are assumed equal
\item \code{FALSE}: component variances are not assumed to be equal
}}
}
\value{
object returned by the \code{\link[=nlm]{nlm()}} function
}
\description{
The \code{\link[=fit_mixture_BinCont]{fit_mixture_BinCont()}} function fits the normal mixture distribution for
a given set of component probabilities. The corresponding component means and
component variances are thus estimated.
}
