% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA.BinCont.R
\name{fit_mixture_BinCont_two_comp}
\alias{fit_mixture_BinCont_two_comp}
\title{Fit two-component normal mixture distribution given component probabilities.}
\usage{
fit_mixture_BinCont_two_comp(
  data_S,
  data_T,
  theta_start_S,
  pi_hier,
  group,
  Diff.Sigma = FALSE
)
}
\arguments{
\item{data_S}{Vector of potential outcomes, either \eqn{S_0} or \eqn{S_1}.}

\item{data_T}{Vector of potential outcomes, either \eqn{T_0} or \eqn{T_1}.}

\item{theta_start_S}{Starting values.}

\item{pi_hier}{Vector of probabilities for the four components:
\eqn{(\pi_{00}, \pi_{01}, \pi_{10}, \pi_{11})}. These are fixed during
estimation.}

\item{group}{For which treatment group are the mixture components estimated:
\itemize{
\item \code{0}: Control group
\item \code{1}: Experimental group
}}

\item{Diff.Sigma}{Are all component variances assumed equal?
\itemize{
\item \code{TRUE}: all component variances within treatment group are assumed equal
\item \code{FALSE}: component variances are not assumed to be equal
}}
}
\value{
object returned by the \code{\link[=nlm]{nlm()}} function
}
\description{
The \code{\link[=fit_mixture_BinCont_two_comp]{fit_mixture_BinCont_two_comp()}} function fits the two-component normal
mixture distribution for a given set of component probabilities. This is an
alternative estimation method where all true endpoint information is taken
into account.
}
\details{
The "original" estimation procedure considers the following mixture
distribution for patient \eqn{i} given the observed data
\deqn{S_0 \sim
 \pi_{00} \cdot N(\mu^{00}_0, \sigma^{00}_0) + \pi_{01} \cdot N(\mu^{01}_0, \sigma^{01}_0) +
 \pi_{10} \cdot N(\mu^{10}_0, \sigma^{10}_0) + \pi_{11} \cdot N(\mu^{11}_0, \sigma^{11}_0)}
for \eqn{Z_{i} = 0}, and analogously for patients with \eqn{Z_{i} = 1}. This
procedure is implemented in the \code{\link[=fit_mixture_BinCont_two_comp]{fit_mixture_BinCont_two_comp()}}.

The estimation procedure that is implemented in this function considers the
following mixture distribution instead:
\deqn{S_0 | T_0 = k \sim
 \pi_{k0} \cdot N(\mu^{k0}_0, \sigma^{k0}_0) + \pi_{k1} \cdot N(\mu^{k1}_0, \sigma^{k1}_0)}
for \eqn{Z_{i} = 0}, and analogously for patients with \eqn{Z_{i} = 1}. This
procedure takes the information in the true endpoint into account.
Consequently, the estimation is simplified from four-component mixtures to
two-component mixtures.
}
