% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{frank_loglik_copula_scale}
\alias{frank_loglik_copula_scale}
\title{Loglikelihood on the Copula Scale for the Frank Copula}
\usage{
frank_loglik_copula_scale(theta, u, v, d1, d2)
}
\arguments{
\item{theta}{Copula parameter}

\item{u}{A numeric vector. Corresponds to first variable on the copula scale.}

\item{v}{A numeric vector. Corresponds to second variable on the copula
scale.}

\item{d1}{An integer vector. Indicates whether first variable is observed or
right-censored,
\itemize{
\item \code{d1[i] = 1} if \code{u[i]} corresponds to non-censored value
\item \code{d1[i] = 0} if \code{u[i]} corresponds to right-censored value
}}

\item{d2}{An integer vector. Indicates whether first variable is observed or
right-censored,
\itemize{
\item \code{d1[i] = 1} if \code{u[i]} corresponds to non-censored value
\item \code{d1[i] = 0} if \code{u[i]} corresponds to right-censored value
}}
}
\value{
Value of the copula loglikelihood evaluated in \code{theta}.
}
\description{
\code{frank_loglik_copula_scale()} computes the loglikelihood on the copula
scale for the Frank copula which is parameterized by \code{theta} as follows:
\deqn{ C(u, v) = - \frac{1}{\theta} \log \left[ 1 - \frac{(1 - e^{-\theta u})(1 - e^{-\theta v})}{1 - e^{-\theta}} \right]}
}
