% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_alpha_ContCont.R
\name{ICA_alpha_ContCont}
\alias{ICA_alpha_ContCont}
\title{Assess surrogacy using a Rényi divergence based family of metrics in the causal-inference single-trial setting in normal case}
\usage{
ICA_alpha_ContCont(
  alpha = numeric(),
  T0S0,
  T1S1,
  T0T0 = 1,
  T1T1 = 1,
  S0S0 = 1,
  S1S1 = 1,
  T0T1 = seq(-1, 1, by = 0.1),
  T0S1 = seq(-1, 1, by = 0.1),
  T1S0 = seq(-1, 1, by = 0.1),
  S0S1 = seq(-1, 1, by = 0.1)
)
}
\arguments{
\item{alpha}{(numeric) is order \verb{alpha in [0, infinity]}}

\item{T0S0}{A scalar or vector that specifies the correlation(s) between the surrogate and the true endpoint in the control treatment condition}

\item{T1S1}{A scalar or vector that specifies the correlation(s) between the surrogate and the true endpoint in the control treatment condition}

\item{T0T0}{A scalar that specifies the variance of the true endpoint in the control treatment condition}

\item{T1T1}{A scalar that specifies the variance of the true endpoint in the control treatment condition}

\item{S0S0}{A scalar that specifies the variance of the true endpoint in the control treatment condition}

\item{S1S1}{A scalar that specifies the variance of the true endpoint in the control treatment condition}

\item{T0T1}{A scalar or vector that contains the correlation(s) between the counterfactuals T0 and T1}

\item{T0S1}{A scalar or vector that contains the correlation(s) between the counterfactuals T0 and S1}

\item{T1S0}{A scalar or vector that contains the correlation(s) between the counterfactuals T1 and S0}

\item{S0S1}{A scalar or vector that contains the correlation(s) between the counterfactuals S0 and S1}
}
\value{
\itemize{
\item Total.Num.Matrices: An object of class numeric that contains the total number of matrices that can be formed as based on the user-specified correlations in the function call.
\item Pos.Def: A data.frame that contains the positive definite matrices that can be formed based on the user-specified correlations. These matrices are used to compute the vector of the \eqn{\rho_{\Delta}} values.
\item rho: A scalar or vector that contains the individual causal association \eqn{\rho_{\Delta}}
\item ICA: A scalar or vector that contains the individual causal association \eqn{\rho_{\Delta}^2=ICA}
\item ICA_alpha: A scalar or vector that contains the individual causal association \eqn{ICA_{\alpha}}
\item Sigmas: A data.frame that contains the \eqn{\sigma_{\Delta T}} and \eqn{\sigma_{\Delta S}}
}
}
\description{
This function is a set of metrics to evaluate surrogacy. \eqn{ICA_{\alpha}} have the similar
mathematical properties with \code{\link[=ICA.ContCont]{ICA.ContCont()}}.
}
\section{New family of surrogacy metrics:}{
ICA was extended as \eqn{ICA_{\alpha}} by proposing a new family of surrogacy metrics derived from Rényi divergence.
Based on this family of divergences, one can propose a more general set of surrogacy metrics defined as:

\deqn{ ICA_{\alpha} = 1 - e^{-2 D_{\alpha}[f(\Delta T,\Delta S), f(\Delta T)f(\Delta S)]} }

when \eqn{\alpha=1} then

\eqn{ICA_{1}=1-e^{-2D_1[f(\Delta T,\Delta S),f(\Delta T)f(\Delta S)]}=1-e^{-2I(\Delta T, \Delta S)}=\rho_{\Delta}^2=ICA}
}

\examples{

ICA_alpha_ContCont(alpha=0.5,
                  T0S0 = 0.9597334, T1S1=0.9644139,
                  T0T0=544.3285, T1T1=550.6597,
                  S0S0=180.6831, S1S1=180.9433,
                  T0T1=-0.9,  T0S1=-0.9,
                  T1S0=-0.9 , S0S1=-0.9)

#More than one alpha value can be calculated simultaneously:

ICA_alpha_ContCont(alpha=c(1.25,1.5,2),
                  T0S0 = 0.9597334, T1S1=0.9644139,
                  T0T0=544.3285, T1T1=550.6597,
                  S0S0=180.6831, S1S1=180.9433,
                  T0T1=-0.9, T0S1=-0.9,
                  T1S0=-0.9, S0S1=-0.9)


}
\references{
Deliorman, G., Alonso, A., & Pardo, M. C. (2025).
A Rényi-Divergence-Based Family of Metrics for the Evaluation of Surrogate Endpoints
in a Causal Inference Framework. \emph{Statistics in Biopharmaceutical Research}.
\url{https://doi.org/10.1080/19466315.2025.2484009}
}
