% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_model_fit.R
\name{make_longitudinal_re_endpoint_specific}
\alias{make_longitudinal_re_endpoint_specific}
\title{Endpoint-specific random effects within treatment arm}
\usage{
make_longitudinal_re_endpoint_specific(
  name,
  S_0 = "none",
  T_0 = "none",
  S_1 = "none",
  T_1 = "none"
)
}
\arguments{
\item{name}{Model name.}

\item{S_0, T_0, S_1, T_1}{Random-effects structure for the surrogate and true
endpoints in the control and treatment arms. One of \code{"none"},
\code{"ri"}, or \code{"rislope"}.}
}
\value{
A random-effects model specification.
}
\description{
Endpoint-specific random effects within treatment arm
}
