% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_model_fit.R
\name{make_longitudinal_re_none}
\alias{make_longitudinal_re_none}
\title{No random effects}
\usage{
make_longitudinal_re_none(name = "Marginal model")
}
\arguments{
\item{name}{Model name.}
}
\value{
A random-effects model specification.
}
\description{
No random effects
}
