% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_BinCont_copula.R
\name{marginal_distribution}
\alias{marginal_distribution}
\title{Fit marginal distribution}
\usage{
marginal_distribution(x, distribution)
}
\arguments{
\item{x}{(numeric) data vector}

\item{distribution}{Distributional family. One of the follwing:
\itemize{
\item \code{"normal"}: normal distribution
\item \verb{"logistic}: logistic distribution as parameterized in \code{dlogis()}
\item \code{"t"}: student t distribution is parameterized in \code{dt()}
\item \code{"lognormal"}: lognormal distribution as parameterized in \code{dlnorm()}
\item \code{"gamma"}: gamma distribution as parameterized in \code{dgamma()}
\item \code{"weibull"}: weibull distribution as parameterized in \code{dweibull()}
}}
}
\value{
Object of class \code{fitdistrplus::fitdist} that represents the
marginal surrogate distribution.
}
\description{
The \code{marginal_distribution()} function is a wrapper for
\code{fitdistrplus::fitdist()} that fits a univariate distribution to a data
vector.
}
