% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{new_vine_copula_fit}
\alias{new_vine_copula_fit}
\title{Constructor for vine copula model}
\usage{
new_vine_copula_fit(fit_0, fit_1, endpoint_types)
}
\arguments{
\item{fit_0}{list returned by \code{\link[=fit_copula_submodel_OrdCont]{fit_copula_submodel_OrdCont()}},
\code{\link[=fit_copula_submodel_ContCont]{fit_copula_submodel_ContCont()}}, or \code{\link[=fit_copula_submodel_OrdOrd]{fit_copula_submodel_OrdOrd()}}.}

\item{fit_1}{list returned by \code{\link[=fit_copula_submodel_OrdCont]{fit_copula_submodel_OrdCont()}},
\code{\link[=fit_copula_submodel_ContCont]{fit_copula_submodel_ContCont()}}, or \code{\link[=fit_copula_submodel_OrdOrd]{fit_copula_submodel_OrdOrd()}}.}

\item{endpoint_types}{Character vector with 2 elements indicating the type of
endpoints. Each element is either \code{"ordinal"} or \code{"continuous"}.}
}
\value{
S3 object of the class \code{vine_copula_fit}.
}
\description{
Constructor for vine copula model
}
\seealso{
\code{\link[=print.vine_copula_fit]{print.vine_copula_fit()}}, \code{\link[=plot.vine_copula_fit]{plot.vine_copula_fit()}}
#should not be used be the user
}
