% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_SurvSurv.R
\name{new_vine_copula_ss_fit}
\alias{new_vine_copula_ss_fit}
\title{Constructor for vine copula model}
\usage{
new_vine_copula_ss_fit(
  fit_0,
  fit_1,
  copula_family,
  knots0,
  knots1,
  knott0,
  knott1,
  copula_rotations
)
}
\arguments{
\item{fit_0}{Estimated parameters in the control group.}

\item{fit_1}{Estimated parameters in the experimental group}

\item{copula_family}{Parametric copula family}

\item{knots0}{placement of knots for Royston-Parmar model}

\item{knots1}{placement of knots for Royston-Parmar model}

\item{knott0}{placement of knots for Royston-Parmar model}

\item{knott1}{placement of knots for Royston-Parmar model}
}
\value{
S3 object
}
\description{
Constructor for vine copula model
}
\examples{
#should not be used be the user
}
