% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalyticSurvCat.R
\name{plot.MetaAnalyticSurvCat}
\alias{plot.MetaAnalyticSurvCat}
\title{Generates a plot of the estimated treatment effects for the surrogate endpoint versus the estimated treatment effects for the true endpoint for an object fitted with the 'MetaAnalyticSurvCat()' function.}
\usage{
\method{plot}{MetaAnalyticSurvCat}(x, ...)
}
\arguments{
\item{x}{An object of class 'MetaAnalyticSurvCat' fitted with the 'MetaAnalyticSurvCat()' function.}

\item{...}{...}
}
\value{
A plot of the type ggplot
}
\description{
Generates a plot of the estimated treatment effects for the surrogate endpoint versus the estimated treatment effects for the true endpoint for an object fitted with the 'MetaAnalyticSurvCat()' function.
}
\examples{
\dontrun{
data("colorectal4")
fit <- MetaAnalyticSurvCat(data = colorectal4, true = truend, trueind = trueind, surrog = surrogend,
               trt = treatn, center = center, trial = trialend, patientid = patid)
plot(fit)
}

}
