% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcat.R
\name{plot.survcat}
\alias{plot.survcat}
\title{Generates a plot of the estimated treatment effects for the surrogate endpoint versus the estimated treatment effects for the true endpoint for an object fitted with the 'survcat()' function.}
\usage{
\method{plot}{survcat}(x, ...)
}
\arguments{
\item{x}{An object of class 'survcat' fitted with the 'survcat()' function.}

\item{...}{...}
}
\value{
A plot of the type ggplot
}
\description{
Generates a plot of the estimated treatment effects for the surrogate endpoint versus the estimated treatment effects for the true endpoint for an object fitted with the 'survcat()' function.
}
\examples{
data("colorectal4")
fit <- survcat(data = colorectal4, true = truend, trueind = trueind, surrog = surrogend,
               trt = treatn, center = center, trial = trialend, patientid = patid)
plot(fit)

}
