% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalyticSurvCat.R
\name{print.MetaAnalyticSurvCat}
\alias{print.MetaAnalyticSurvCat}
\title{Prints all the elements of an object fitted with the 'MetaAnalyticSurvCat()' function.}
\usage{
\method{print}{MetaAnalyticSurvCat}(x, ...)
}
\arguments{
\item{x}{An object of class 'MetaAnalyticSurvCat' fitted with the 'MetaAnalyticSurvCat()' function.}

\item{...}{...}
}
\value{
The surrogacy measures with their 95\% confidence intervals and the estimated treatment effect on the surrogate and true endpoint.
}
\description{
Prints all the elements of an object fitted with the 'MetaAnalyticSurvCat()' function.
}
\examples{
\dontrun{
data("colorectal4")
fit <- MetaAnalyticSurvCat(data = colorectal4, true = truend, trueind = trueind, surrog = surrogend,
                           trt = treatn, center = center, trial = trialend, patientid = patid,
                           adjustment="unadjusted")
print(fit)
}
}
