% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalyticSurvSurv.R
\name{print.MetaAnalyticSurvSurv}
\alias{print.MetaAnalyticSurvSurv}
\title{Prints all the elements of an object fitted with the 'MetaAnalyticSurvSurv()' function.}
\usage{
\method{print}{MetaAnalyticSurvSurv}(x, ...)
}
\arguments{
\item{x}{An object of class 'MetaAnalyticSurvSurv' fitted with the 'MetaAnalyticSurvSurv()' function.}

\item{...}{...}
}
\value{
The surrogacy measures with their 95\% confidence intervals and the estimated treatment effect on the surrogate and true endpoint.
}
\description{
Prints all the elements of an object fitted with the 'MetaAnalyticSurvSurv()' function.
}
\examples{
\dontrun{
data("colorectal4")
fit <- MetaAnalyticSurvSurv(data=Ovarian,true=Surv,trueind=SurvInd,surrog=Pfs,surrogind=PfsInd,
                            trt=Treat,center=Center,trial=Center,patientid=Patient,
                            copula="Plackett",adjustment="unadjusted")
print(fit)
}
}
