% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcat.R
\name{print.survcat}
\alias{print.survcat}
\title{Prints all the elements of an object fitted with the 'survcat()' function.}
\usage{
\method{print}{survcat}(x, ...)
}
\arguments{
\item{x}{An object of class 'survcat' fitted with the 'survcat()' function.}

\item{...}{...}
}
\value{
The surrogacy measures with their 95\% confidence intervals and the estimated treament effect on the surrogate and true endpoint.
}
\description{
Prints all the elements of an object fitted with the 'survcat()' function.
}
\examples{
\dontrun{
data("colorectal4")
fit <- survcat(data = colorectal4, true = truend, trueind = trueind, surrog = surrogend,
               trt = treatn, center = center, trial = trialend, patientid = patid)
print(fit)
}
}
