% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{print.vine_copula_fit}
\alias{print.vine_copula_fit}
\title{Print summary of fitted copula model}
\usage{
\method{print}{vine_copula_fit}(x, ...)
}
\arguments{
\item{x}{Fitted-model object returned by \code{\link[=fit_copula_ContCont]{fit_copula_ContCont()}},
\code{\link[=fit_copula_OrdCont]{fit_copula_OrdCont()}}, or \code{\link[=fit_copula_OrdOrd]{fit_copula_OrdOrd()}}.}

\item{...}{not used}
}
\description{
Print summary of fitted copula model
}
