% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis_BinCont_copula.R
\name{sensitivity_analysis_BinCont_copula}
\alias{sensitivity_analysis_BinCont_copula}
\title{Perform Sensitivity Analysis for the Individual Causal Association with a
Continuous Surrogate and Binary True Endpoint}
\usage{
sensitivity_analysis_BinCont_copula(
  fitted_model,
  n_sim,
  cond_ind,
  lower = c(-1, -1, -1, -1),
  upper = c(1, 1, 1, 1),
  marg_association = TRUE,
  n_prec = 10000,
  ncores = 1
)
}
\arguments{
\item{fitted_model}{Returned value from \code{\link[=fit_model_BinCont]{fit_model_BinCont()}}. This object
contains the estimated identifiable part of the joint distribution for the
potential outcomes.}

\item{n_sim}{Number of copula parameter vectors to be sampled.}

\item{cond_ind}{(boolean) Indicates whether conditional independence is
assumed, see Conditional Independence. Defaults to \code{FALSE}.}

\item{lower}{(numeric) Vector of length 4 that provides the lower limit,
\eqn{\boldsymbol{a} = (a_{23}, a_{13;2}, a_{24;3},
  a_{14;23})'}. Defaults to \code{c(-1, -1, -1, -1)}. If the provided lower limit
is smaller than what is allowed for a particular copula family, then the
copula family's lowest possible value is used instead.}

\item{upper}{(numeric) Vector of length 4 that provides the upper limit,
\eqn{\boldsymbol{b} = (b_{23}, b_{13;2}, b_{24;3},
  b_{14;23})'}. Defaults to \code{c(1, 1, 1, 1)}.}
}
\value{
A data frame is returned. Each row represents one replication in the
sensitivity analysis. The returned data frame always contains the following
columns:
\itemize{
\item \code{R2H}, \code{sp_rho}, \code{minfo}: ICA as quantified by \eqn{R^2_H}, Spearman's rho, and
Kendall's tau, respectively.
\item \code{c12}, \code{c34}: estimated copula parameters.
\item \code{c23}, \code{c13_2}, \code{c24_3}, \code{c14_23}: sampled copula parameters of the
unidentifiable copulas in the D-vine copula. The parameters correspond to
the parameterization of the \code{copula_family2} copula as in the \code{copula}
R-package.
\item \code{r12}, \code{r34}: Fixed rotation parameters for the two identifiable copulas.
\item \code{r23}, \code{r13_2}, \code{r24_3}, \code{r14_23}: Sampled rotation parameters of the
unidentifiable copulas in the D-vine copula. These values are constant for
the Gaussian copula family since that copula is invariant to rotations.
}

The returned data frame also contains the following columns when
\code{marg_association} is \code{TRUE}:
\itemize{
\item \code{sp_s0s1}, \code{sp_s0t0}, \code{sp_s0t1}, \code{sp_s1t0}, \code{sp_s1t1}, \code{sp_t0t1}:
Spearman's rho between the corresponding potential outcomes. Note that these
associations refer to the observable potential outcomes. In contrary, the
estimated association parameters from \code{\link[=fit_model_BinCont]{fit_model_BinCont()}} refer to
associations on a latent scale.
}
}
\description{
Perform Sensitivity Analysis for the Individual Causal Association with a
Continuous Surrogate and Binary True Endpoint
}
\section{Quantifying Surrogacy}{
In the information-theoretic causal-inference (ITCI) framework to evaluate
surrogate endpoints, the ICA is the measure of primary interest. This measure
quantifies how much information the individual causal treatment effect on the
surrogate (\eqn{\Delta S}) provides on the individual causal treatment effect
on the true endpoint (\eqn{\Delta T}). The mutual information between
\eqn{\Delta S} and \eqn{\Delta T}, denoted by \eqn{I(\Delta S; \Delta T)} is
a natural candidate to quantify this amount of shared information. However,
the mutual information is difficult to interpret as there does not exist a
general upper bound. Alonso et al. (na) therfore proposed to quantify the ICA
thorugh a transformation of the mutual information that is guaranteed to lie
in the unit interval. It is the following measure, \deqn{R^2_H =
\frac{I(\Delta S; \Delta T)}{H(\Delta T)}} where \eqn{H(\Delta T)} is the
entropy of \eqn{\Delta T}. By token of that transformation of the mutual
information, \eqn{R^2_H} is restricted to the unit interval where 0 indicates
independence, and 1 a functional relationship between \eqn{\Delta S} and
\eqn{\Delta T}.

The association between \eqn{\Delta S} and \eqn{\Delta T} can also be
quantified by Spearman's \eqn{\rho} (or Kendall's \eqn{\tau}). This quantity
requires appreciably less computing time than the mutual information. This
quantity is therefore always returned for every replication of the
sensitivity analysis.
}

