% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_intervals.R
\name{sensitivity_intervals_Dvine}
\alias{sensitivity_intervals_Dvine}
\title{Compute Sensitivity Intervals}
\usage{
sensitivity_intervals_Dvine(
  fitted_model,
  sens_results,
  measure = "ICA",
  B = 200,
  alpha = 0.05,
  n_prec = 5000
)
}
\arguments{
\item{fitted_model}{Returned value from \code{\link[=fit_model_SurvSurv]{fit_model_SurvSurv()}}. This object
contains the estimated identifiable part of the joint distribution for the
potential outcomes.}

\item{sens_results}{Dataframe returned by
\code{\link[=sensitivity_analysis_SurvSurv_copula]{sensitivity_analysis_SurvSurv_copula()}}. If additional assumptions need to
be incorporated, this dataframe can first be filtered.}

\item{measure}{Compute intervals for which measure of surrogacy? Defaults to
\code{"ICA"}. See first column names of \code{sens_results} for other possibilities.}

\item{B}{Number of bootstrap replications}

\item{alpha}{(numeric) \code{1 - alpha} is the level of the confidence interval}

\item{n_prec}{Number of Monte Carlo samples for the computation of the mutual
information.}
}
\value{
An S3 object of the class \code{sensitivity_intervals_Dvine} which can be
printed.
}
\description{
\code{\link[=sensitivity_intervals_Dvine]{sensitivity_intervals_Dvine()}} computes the intervals of ignorance and
uncertainty within the information-theoretic causal inference framework when
the data are modeled with a D-vine copula model.
}
\examples{
# Load Ovarian data
data("Ovarian")
# Recode the Ovarian data in the semi-competing risks format.
data_scr = data.frame(
  ttp = Ovarian$Pfs,
  os = Ovarian$Surv,
  treat = Ovarian$Treat,
  ttp_ind = ifelse(
    Ovarian$Pfs == Ovarian$Surv &
      Ovarian$SurvInd == 1,
    0,
    Ovarian$PfsInd
  ),
  os_ind = Ovarian$SurvInd
)
# Fit copula model.
fitted_model = fit_model_SurvSurv(data = data_scr,
                                  copula_family = "clayton",
                                  n_knots = 1)
# Illustration with small number of replications and low precision
sens_results = sensitivity_analysis_SurvSurv_copula(fitted_model,
                  n_sim = 5,
                  n_prec = 2000,
                  copula_family2 = "clayton",
                  cond_ind = TRUE)
# Compute intervals of ignorance and uncertainty. Again, the number of
# bootstrap replications should be larger in practice.
sensitivity_intervals_Dvine(fitted_model, sens_results, B = 50)
}
