% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalyticSurvCat.R
\name{summary.MetaAnalyticSurvCat}
\alias{summary.MetaAnalyticSurvCat}
\title{Provides a summary of the surrogacy measures for an object fitted with the 'MetaAnalyticSurvCat()' function.}
\usage{
\method{summary}{MetaAnalyticSurvCat}(object, ...)
}
\arguments{
\item{object}{An object of class 'MetaAnalyticSurvCat' fitted with the 'MetaAnalyticSurvCat()' function.}

\item{...}{...}
}
\value{
The surrogacy measures with their 95\% confidence intervals.
}
\description{
Provides a summary of the surrogacy measures for an object fitted with the 'MetaAnalyticSurvCat()' function.
}
\examples{
\dontrun{
data("colorectal4")
fit <- MetaAnalyticSurvCat(data = colorectal4, true = truend, trueind = trueind, surrog = surrogend,
                           trt = treatn, center = center, trial = trialend, patientid = patid,
                           adjustment="unadjusted")
summary(fit)
}
}
