% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surrogate-package.R
\docType{package}
\name{Surrogate-package}
\alias{Surrogate}
\alias{Surrogate-package}
\title{Surrogate: Evaluation of Surrogate Endpoints in Clinical Trials}
\description{
In a clinical trial, it frequently occurs that the most credible outcome to evaluate the effectiveness of a new therapy (the true endpoint) is difficult to measure. In such a situation, it can be an effective strategy to replace the true endpoint by a (bio)marker that is easier to measure and that allows for a prediction of the treatment effect on the true endpoint (a surrogate endpoint). The package 'Surrogate' allows for an evaluation of the appropriateness of a candidate surrogate endpoint based on the meta-analytic, information-theoretic, and causal-inference frameworks. Part of this software has been developed using funding provided from the European Union's Seventh Framework Programme for research, technological development and demonstration (Grant Agreement no 602552), the Special Research Fund (BOF) of Hasselt University (BOF-number: BOF2OCPO3), GlaxoSmithKline Biologicals, Baekeland Mandaat (HBC.2022.0145), and Johnson & Johnson Innovative Medicine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/florianstijven/Surrogate-development}
  \item Report bugs at \url{https://github.com/florianstijven/Surrogate-development/issues}
}

}
\author{
\strong{Maintainer}: Wim Van Der Elst \email{wim.vanderelst@gmail.com}

Authors:
\itemize{
  \item Florian Stijven \email{florian.stijven@kuleuven.be}
  \item Fenny Ong
  \item Dries De Witte
  \item Paul Meyvisch
  \item Alvaro Poveda
  \item Ariel Alonso
  \item Hannah Ensor
  \item Christoper Weir
  \item Geert Molenberghs
}

}
\keyword{internal}
