% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_SurvSurv.R
\name{survival_survival_loglik}
\alias{survival_survival_loglik}
\title{Loglikelihood function for survival-continuous copula model}
\usage{
survival_survival_loglik(
  para,
  X,
  delta_X,
  Y,
  delta_Y,
  copula_family,
  knotsx,
  knotsy
)
}
\description{
Loglikelihood function for survival-continuous copula model
}
