% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_SurvSurv.R
\name{SurvSurv_starting_values}
\alias{SurvSurv_starting_values}
\title{Title}
\usage{
SurvSurv_starting_values(X, delta_X, Y, delta_Y, copula_family, nknots)
}
\arguments{
\item{nknots}{}
}
\description{
Title
}
