% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_copula_models.R
\name{association_gof_copula}
\alias{association_gof_copula}
\title{Produce Associational GoF plot}
\usage{
association_gof_copula(fitted_submodel, treat, endpoint_types)
}
\arguments{
\item{fitted_submodel}{List returned by \code{\link[=fit_copula_submodel_OrdCont]{fit_copula_submodel_OrdCont()}},
\code{\link[=fit_copula_submodel_ContCont]{fit_copula_submodel_ContCont()}}, or \code{\link[=fit_copula_submodel_OrdOrd]{fit_copula_submodel_OrdOrd()}}.}

\item{treat}{Value for the treatment indicator.}

\item{endpoint_types}{Character vector with 2 elements indicating the type of endpoints. Each element is either
\code{"ordinal"} or \code{"continuous"}.}
}
\description{
Produce Associational GoF plot
}
\seealso{
\code{\link[=plot.vine_copula_fit]{plot.vine_copula_fit()}}
}
