% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_BinCont_copula.R
\name{binary_continuous_loglik}
\alias{binary_continuous_loglik}
\title{Loglikelihood function for binary-continuous copula model}
\usage{
binary_continuous_loglik(para, X, Y, copula_family, marginal_surrogate)
}
\arguments{
\item{para}{Parameter vector}

\item{X}{First variable (continuous)}

\item{Y}{Second variable (binary, $0$ or $1$)}

\item{marginal_surrogate}{Marginal distribution for the surrogate}
}
\description{
Loglikelihood function for binary-continuous copula model
}
