% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{estimate_marginal}
\alias{estimate_marginal}
\title{Estimate marginal distribution using ML}
\usage{
estimate_marginal(Y, marginal_Y, starting_values)
}
\arguments{
\item{Y}{Observations (continuous)}

\item{marginal_Y}{List with the following five elements (in order):
\itemize{
\item Density function with first argument \code{x} and second argument \code{para} the parameter
vector for this distribution.
\item Distribution function with first argument \code{x} and second argument \code{para}.
\item Inverse distribution function with first argument \code{p} and second argument \code{para}.
\item The number of elements in \code{para}.
\item Starting values for \code{para}.
}}

\item{starting_values}{Starting values for \code{marginal_Y}}
}
\value{
Estimated parameters
}
\description{
\code{\link[=estimate_marginal]{estimate_marginal()}} estimates the marginal distribution specified by
\code{marginal_Y} using maximum likelihood. The optimizer is Newton-Raphson.
}
