% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_BinCont_copula.R
\name{fit_copula_model_BinCont}
\alias{fit_copula_model_BinCont}
\title{Fit copula model for binary true endpoint and continuous surrogate endpoint}
\usage{
fit_copula_model_BinCont(
  data,
  copula_family,
  marginal_surrogate,
  fitted_model = NULL,
  hessian = TRUE,
  maxit = 500
)
}
\arguments{
\item{data}{A data frame in the correct format (See details).}

\item{copula_family}{One of the following parametric copula families:
\code{"clayton"}, \code{"frank"}, \code{"gaussian"}, or \code{"gumbel"}.}

\item{fitted_model}{Fitted model from which initial values are extracted. If
\code{NULL} (default), standard initial values are used. This option intended
for when a model is repeatedly fitted, e.g., in a bootstrap.}

\item{hessian}{A boolean.
\itemize{
\item \code{TRUE} (default): Hessian is computed
\item \code{FALSE}: Hessian is not computed. This can save a small amount of time. This
can be useful when a model is repeatedly fitted, e.g., in a bootstrap.
}}

\item{maxit}{Maximum number of iterations for the numeric optimization,
defaults to 500.}
}
\description{
The function \code{fit_copula_model_BinCont()} fits the copula model for a
continuous surrogate endpoint and binary true endpoint. Because the bivariate
distributions of the surrogate-true endpoint pairs are functionally
independent across treatment groups, a bivariate distribution is fitted in
each treatment group separately.
}
