% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_ContCont_copula.R
\name{fit_copula_submodel_ContCont}
\alias{fit_copula_submodel_ContCont}
\title{Fit ordinal-continuous copula submodel}
\usage{
fit_copula_submodel_ContCont(
  X,
  Y,
  copula_family,
  marginal_X,
  marginal_Y,
  start_X,
  start_Y,
  start_copula,
  method = "BFGS",
  names_XY = c("Surr", "True"),
  twostep = FALSE,
  copula_transform = function(x) x,
  ...
)
}
\arguments{
\item{X}{First variable (Continuous)}

\item{Y}{Second variable (Continuous)}

\item{copula_family}{Copula family, one of the following:
\itemize{
\item \code{"clayton"}
\item \code{"frank"}
\item \code{"gumbel"}
\item \code{"gaussian"}
}}

\item{marginal_X}{List with the following three elements (in order):
\itemize{
\item Density function with first argument \code{x} and second argument \code{para} the parameter
vector for this distribution.
\item Distribution function with first argument \code{x} and second argument \code{para} the parameter
vector for this distribution.
\item The number of elements in \code{para}.
}}

\item{marginal_Y}{See \code{marginal_X}}

\item{name}{description}
}
\value{
A list with four elements:
\itemize{
\item ml_fit: object of class \code{maxLik::maxLik} that contains the estimated copula
model.
\item marginal_X: list with the estimated cdf, pdf, and inverse cdf for X.
\item marginal_Y: list with the estimated cdf, pdf, and inverse cdf for X.
\item copula_family: string that indicates the copula family
}
}
\description{
The \code{fit_copula_submodel_ContCont()} function fits the copula (sub)model for a
continuous surrogate and true endpoint with maximum likelihood.
}
