% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_OrdCont_copula.R
\name{fit_copula_submodel_OrdCont}
\alias{fit_copula_submodel_OrdCont}
\title{Fit ordinal-continuous copula submodel}
\usage{
fit_copula_submodel_OrdCont(
  X,
  Y,
  copula_family,
  marginal_Y,
  start_Y,
  start_copula,
  method = "BFGS",
  K,
  names_XY = c("Surr", "True"),
  twostep = FALSE
)
}
\arguments{
\item{X}{First variable (Ordinal with \eqn{K} categories)}

\item{Y}{Second variable (Continuous)}

\item{copula_family}{Copula family, one of the following:
\itemize{
\item \code{"clayton"}
\item \code{"frank"}
\item \code{"gumbel"}
\item \code{"gaussian"}
}}

\item{marginal_Y}{List with the following five elements (in order):
\itemize{
\item Density function with first argument \code{x} and second argument \code{para} the parameter
vector for this distribution.
\item Distribution function with first argument \code{x} and second argument \code{para}.
\item Inverse distribution function with first argument \code{p} and second argument \code{para}.
\item The number of elements in \code{para}.
\item Starting values for \code{para}.
}}

\item{K}{Number of categories in \code{X}.}

\item{name}{description}
}
\value{
A list with four elements:
\itemize{
\item ml_fit: object of class \code{maxLik::maxLik} that contains the estimated copula
model.
\item marginal_X: list with the estimated cdf, pdf, and inverse cdf for X.
\item marginal_Y: list with the estimated cdf, pdf, and inverse cdf for X.
\item copula_family: string that indicates the copula family
}
}
\description{
The \code{fit_copula_submodel_OrdCont()} function fits the copula (sub)model for a
continuous surrogate and an ordinal true endpoint with maximum likelihood.
}
