% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{loglik_copula_scale}
\alias{loglik_copula_scale}
\title{Loglikelihood on the Copula Scale}
\usage{
loglik_copula_scale(theta, u, v, d1, d2, copula)
}
\arguments{
\item{theta}{Copula parameter}

\item{u}{A numeric vector. Corresponds to first variable on the copula scale.}

\item{v}{A numeric vector. Corresponds to second variable on the copula
scale.}

\item{d1}{An integer vector. Indicates whether first variable is observed or
right-censored,
\itemize{
\item \code{d1[i] = 1} if \code{u[i]} corresponds to non-censored value
\item \code{d1[i] = 0} if \code{u[i]} corresponds to right-censored value
}}

\item{d2}{An integer vector. Indicates whether first variable is observed or
right-censored,
\itemize{
\item \code{d1[i] = 1} if \code{u[i]} corresponds to non-censored value
\item \code{d1[i] = 0} if \code{u[i]} corresponds to right-censored value
}}

\item{copula}{Copula family, one of the following:
\itemize{
\item \code{"clayton"}
\item \code{"frank"}
\item \code{"gumbel"}
\item \code{"gaussian"}

The parameterization of the respective copula families can be found in the
help files of the dedicated functions named \code{copula_loglik_copula_scale()}.
}}
}
\value{
Value of the copula loglikelihood evaluated in \code{theta}.
}
\description{
\code{loglik_copula_scale()} computes the loglikelihood on the copula scale for
possibly right-censored data.
}
