% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_copula_models.R
\name{marginal_gof_copula}
\alias{marginal_gof_copula}
\title{Produce marginal GoF plot}
\usage{
marginal_gof_copula(marginal, observed, name, type, treat)
}
\arguments{
\item{marginal}{Estimated marginal distribution represented by a list with
three elements in the following order: the estimated cdf, pdf, and inverse
cdf.}

\item{observed}{Observed values. These are used for the histogram.}

\item{name}{Name of the endpoint (used in the plot title).}

\item{type}{Type of endpoint: \code{"ordinal"} or \code{"continuous"}}

\item{treat}{Value for the treatment indicator.}
}
\description{
Produce marginal GoF plot
}
\seealso{
\code{\link[=plot.vine_copula_fit]{plot.vine_copula_fit()}}
}
