% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{new_vine_copula_fit}
\alias{new_vine_copula_fit}
\title{Constructor for vine copula model}
\usage{
new_vine_copula_fit(fit_0, fit_1, endpoint_types)
}
\arguments{
\item{fit_0}{list returned by \code{\link[=fit_copula_submodel_OrdCont]{fit_copula_submodel_OrdCont()}},
\code{\link[=fit_copula_submodel_ContCont]{fit_copula_submodel_ContCont()}}, or \code{\link[=fit_copula_submodel_OrdOrd]{fit_copula_submodel_OrdOrd()}}.}

\item{fit_1}{list returned by \code{\link[=fit_copula_submodel_OrdCont]{fit_copula_submodel_OrdCont()}},
\code{\link[=fit_copula_submodel_ContCont]{fit_copula_submodel_ContCont()}}, or \code{\link[=fit_copula_submodel_OrdOrd]{fit_copula_submodel_OrdOrd()}}.}
}
\value{
S3 object
}
\description{
Constructor for vine copula model
}
\examples{
#should not be used be the user
}
