% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_BinCont_copula.R
\name{sample_deltas_BinCont}
\alias{sample_deltas_BinCont}
\title{Sample individual casual treatment effects from given D-vine copula model in
binary continuous setting}
\usage{
sample_deltas_BinCont(
  copula_par,
  rotation_par,
  copula_family1,
  copula_family2 = copula_family1,
  n,
  q_S0,
  q_S1
)
}
\arguments{
\item{copula_par}{Parameter vector for the sequence of bivariate copulas that
define the D-vine copula. The elements of \code{copula_par} correspond to
\eqn{(c_{12}, c_{23}, c_{34}, c_{13;2}, c_{24;3}, c_{14;23})}.}

\item{rotation_par}{Vector of rotation parameters for the sequence of
bivariate copulas that define the D-vine copula. The elements of
\code{rotation_par} correspond to \eqn{(c_{12}, c_{23}, c_{34}, c_{13;2},
  c_{24;3}, c_{14;23})}.}

\item{copula_family1}{Copula family of \eqn{c_{12}} and \eqn{c_{34}}. For the
possible options, see \code{loglik_copula_scale()}.}

\item{copula_family2}{Copula family of the other bivariate copulas. For the
possible options, see \code{loglik_copula_scale()}.}

\item{n}{Number of samples to be taken from the D-vine copula.}

\item{q_S0}{Quantile function for the distribution of \eqn{S_0}.}

\item{q_S1}{Quantile function for the distribution of \eqn{S_1}.}
}
\value{
A data frame
}
\description{
Sample individual casual treatment effects from given D-vine copula model in
binary continuous setting
}
