\name{summary}
\alias{summary}
\alias{summary.BifixedContCont}
\alias{summary.FixedContContIT}
\alias{summary.BimixedContCont}
\alias{summary.UnifixedContCont}
\alias{summary.UnimixedContCont}
\alias{summary.ICA.ContCont}
\alias{summary.MICA.ContCont}
\alias{summary.MinSurrContCont}
\alias{summary.MixedContContIT}
\alias{summary.Single.Trial.RE.AA}
\alias{summary.SPF.BinBin}
\alias{summary.ICA.BinBin}
\alias{summary.TrialLevelContCont}
\alias{summary.TwoStageSurvSurv}
\title{Summary}
\description{summary}
\usage{
\method{summary}{BifixedContCont}(object, ..., Object)
\method{summary}{BimixedContCont}(object, ..., Object)
\method{summary}{UnifixedContCont}(object, ..., Object)
\method{summary}{UnimixedContCont}(object, ..., Object)
\method{summary}{FixedContContIT}(object, ..., Object)
\method{summary}{ICA.ContCont}(object, ..., Object)
\method{summary}{MICA.ContCont}(object, ..., Object)
\method{summary}{MinSurrContCont}(object, ..., Object)
\method{summary}{MixedContContIT}(object, ..., Object)
\method{summary}{Single.Trial.RE.AA}(object, ..., Object)
\method{summary}{SPF.BinBin}(object, ..., Object)
\method{summary}{ICA.BinBin}(object, ..., Object)
\method{summary}{TrialLevelContCont}(object, ..., Object)
\method{summary}{TwoStageSurvSurv}(object, ..., Object)
}
\keyword{internal}
