\name{summary}
\alias{summary}
\alias{summary.BifixedContCont}
\alias{summary.FixedContContIT}
\alias{summary.BimixedContCont}
\alias{summary.UnifixedContCont}
\alias{summary.UnimixedContCont}
\alias{summary.ICA.ContCont}
\alias{summary.MICA.ContCont}
\alias{summary.MinSurrContCont}
\alias{summary.MixedContContIT}
\alias{summary.Single.Trial.RE.AA}
\alias{summary.SPF.BinBin}
\alias{summary.ICA.BinBin}
\alias{summary.TrialLevelMA}
\alias{summary.TwoStageSurvSurv}
\alias{summary.Prentice}
\alias{summary.PredTrialTContCont}
\alias{summary.FixedBinBinIT}
\alias{summary.FixedBinContIT}
\alias{summary.FixedContBinIT}
\alias{summary.SurvSurv}
\alias{summary.TrialLevelIT}
\alias{summary.MaxEntICA.BinBin}
\alias{summary.MaxEntSPF.BinBin}
\alias{summary.ICA.BinCont}
\title{Summary}
\description{summary}
\usage{
\method{summary}{BifixedContCont}(object, ..., Object)
\method{summary}{BimixedContCont}(object, ..., Object)
\method{summary}{UnifixedContCont}(object, ..., Object)
\method{summary}{UnimixedContCont}(object, ..., Object)
\method{summary}{FixedContContIT}(object, ..., Object)
\method{summary}{ICA.ContCont}(object, ..., Object)
\method{summary}{MICA.ContCont}(object, ..., Object)
\method{summary}{MinSurrContCont}(object, ..., Object)
\method{summary}{MixedContContIT}(object, ..., Object)
\method{summary}{Single.Trial.RE.AA}(object, ..., Object)
\method{summary}{SPF.BinBin}(object, ..., Object)
\method{summary}{ICA.BinBin}(object, ..., Object)
\method{summary}{TrialLevelMA}(object, ..., Object)
\method{summary}{TwoStageSurvSurv}(object, ..., Object)
\method{summary}{Prentice}(object, ..., Object)
\method{summary}{PredTrialTContCont}(object, ..., Object)
\method{summary}{FixedBinBinIT}(object, ..., Object)
\method{summary}{FixedBinContIT}(object, ..., Object)
\method{summary}{FixedContBinIT}(object, ..., Object)
\method{summary}{SurvSurv}(object, ..., Object)
\method{summary}{TrialLevelIT}(object, ..., Object)
\method{summary}{MaxEntICA.BinBin}(object, ..., Object)
\method{summary}{MaxEntSPF.BinBin}(object, ..., Object)
\method{summary}{ICA.BinCont}(object, ..., Object)
}
\keyword{internal}